refactor: remove .component from component names
