+++ src/components/common/bbImage.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement, ComponentType, JSX, FC } from "react";
-import Skeleton from "./skeleton.component";
+import Skeleton from "./skeleton";
 
 type ImageProps = JSX.IntrinsicElements["img"];
 type AsComponent = ("img" | "image") | ComponentType<Partial<ImageProps>>;
