+++ src/components/common/widgets/widget.tsx
