+++ src/pages/sonic._index.tsx
@@ -1,4 +1,4 @@
-import Widget from "../components/common/widgets/widget.component";
+import Widget from "../components/common/widgets/widget";
 
 const Sonic: React.FC = () => {
   return (
