+++ src/pages/user.auth.logout.tsx
@@ -1,4 +1,4 @@
-import Widget from "../components/common/widgets/widget.component";
+import Widget from "../components/common/widgets/widget";
 
 const UserLogout: React.FC = () => {
   return (
