+++ src/pages/user.profile.$userId.tsx
@@ -1,9 +1,9 @@
 import { useBBQuery } from "../hooks/useBBQuery";
 import type { User } from "../types/user";
-import UserLeftPane from "../components/user/userLeftPane.component";
-import Accordion from "../components/common/accordion/Accordion.component";
+import UserLeftPane from "../components/user/userLeftPane";
+import Accordion from "../components/common/accordion/Accordion";
 import parse from "html-react-parser/lib/index";
-import BBInput from "../components/common/forms/BBInput.component";
+import BBInput from "../components/common/forms/BBInput";
 
 const UserProfileMaster: React.FC = () => {
   const { userId } = useParams();
