+++ src/root.tsx
@@ -1,7 +1,7 @@
 import "./assets/App.css";
 import UserProvider from "./providers/user/userProvider";
 import QueryProvider from "./providers/query/queryProvider";
-import RootLayout from "./rootLayout.component";
+import RootLayout from "./root.layout";
 
 const TanStackQueryDevtools = import.meta.env.DEV
   ? lazy(() =>
